types/batchAccount file created
